Remove unused usings and move CommandFieldInfoExtractor to Internal namespace
Clean up unused using directives across all command handlers and move
CommandFieldInfoExtractor to UniCli.Server.Editor.Internal namespace.
Update Rider DotSettings to suppress namespace mismatch warnings for
handler subdirectories.
