+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/AssemblyDefinition/AssemblyDefinitionCreateHandler.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
-using UniCli.Server.Editor;
 using UnityEditor;
 using UnityEditor.Compilation;
 using UnityEngine;
