+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/AssemblyDefinition/AssemblyDefinitionGetHandler.cs
@@ -2,7 +2,6 @@ using System;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using UniCli.Server.Editor;
 using UnityEditor;
 using UnityEditor.Compilation;
 using UnityEngine;
