+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/AssemblyDefinition/AssemblyDefinitionListHandler.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using UniCli.Protocol;
-using UniCli.Server.Editor;
 using UnityEditor.Compilation;
 using UnityEngine;
