+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/CompileHandler.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using UniCli.Protocol;
-using UniCli.Server.Editor;
 using UnityEditor;
 using UnityEditor.Compilation;
