+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/CompilePlayerHandler.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
-using UniCli.Protocol;
 using UnityEditor;
 using UnityEditor.Build.Player;
 using UnityEditor.Compilation;
