+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/Console/EditorLogHandler.cs
@@ -1,7 +1,5 @@
 using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
-using UniCli.Protocol;
 
 namespace UniCli.Server.Editor.Handlers
 {
