+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/ICommandHandler.cs
@@ -1,7 +1,6 @@
-using System;
 using System.Threading.Tasks;
 using UniCli.Protocol;
-using UniCli.Server.Editor;
+using UniCli.Server.Editor.Internal;
 using UnityEngine;
 
 namespace UniCli.Server.Editor.Handlers
