+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/Menu/ExecuteMenuItemHandler.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Threading.Tasks;
-using UniCli.Protocol;
 using UnityEditor;
 
 namespace UniCli.Server.Editor.Handlers
