+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/Menu/GetMenuItemsHandler.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Threading.Tasks;
-using UniCli.Protocol;
 using UnityEditor;
 
 namespace UniCli.Server.Editor.Handlers
