+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/PackageManager/PackageManagerAddHandler.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Threading.Tasks;
-using UniCli.Server.Editor;
 using UnityEditor.PackageManager;
 
 namespace UniCli.Server.Editor.Handlers
