+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/PackageManager/PackageManagerListHandler.cs
@@ -2,10 +2,7 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 using UniCli.Protocol;
-using UniCli.Server.Editor;
-using UnityEditor;
 using UnityEditor.PackageManager;
-using UnityEditor.PackageManager.Requests;
 
 namespace UniCli.Server.Editor.Handlers
 {
