+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/Prefab/PrefabSaveHandler.cs
@@ -1,8 +1,6 @@
 using System;
 using System.Threading.Tasks;
-using UniCli.Protocol;
 using UnityEditor;
-using UnityEngine;
 
 namespace UniCli.Server.Editor.Handlers
 {
