+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/TestRunner/TestRunnerHandler.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Threading.Tasks;
 using UniCli.Protocol;
-using UniCli.Server.Editor;
 using UnityEditor.TestTools.TestRunner.Api;
 using UnityEngine;
