+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/TypeCacheListHandler.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Linq;
 using System.Threading.Tasks;
-using UniCli.Protocol;
 using UnityEditor;
 
 namespace UniCli.Server.Editor.Handlers
