+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/TypeInspectHandler.cs
@@ -2,7 +2,6 @@ using System;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
-using UniCli.Protocol;
 
 namespace UniCli.Server.Editor.Handlers
 {
