+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/UnitySearchHandler.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using UniCli.Protocol;
 using UnityEditor.Search;
 
 namespace UniCli.Server.Editor.Handlers
